requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
setuptools==78.1.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
wheel==0.45.1
